backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
dj-database-url

pytest
pytest-django
pytest-cov
factory-boy
faker

drf-spectacular
python-dotenv
